Add more capability to slice_assign (#7)
