+++ candle-core/src/lib.rs
@@ -69,6 +69,7 @@ pub mod streaming;
 mod strided_index;
 mod tensor;
 mod tensor_cat;
+mod tensor_indexing;
 pub mod test_utils;
 pub mod utils;
 mod variable;
